Change unit tests into integration tests
